refactor: scroll action to ContentLayout
