+++ src/web/src/components/common/index.ts
@@ -1,7 +1,6 @@
 export * from './Modal';
 export { default as BottomSheet } from './BottomSheetModal';
 export { default as Button } from './Button';
-export { default as ContentLayout } from './ContentLayout';
 export { default as Icon } from './Icon';
 export { default as Input } from './Input';
 export { default as Tabs } from './Tabs';
