+++ src/web/src/components/index.ts
@@ -2,6 +2,7 @@ export * from './common';
 export { default as AccessibleIconButton } from './AccessibleIconButton';
 export { default as AsyncBoundary } from './AsyncBoundary';
 export { default as BottomNavigation } from './BottomNavigation';
+export { default as ContentLayout } from './ContentLayout';
 export { default as FloatingButton } from './FloatingButton';
 export { default as Header } from './Header';
 export { default as MenuModal } from './MenuModal';
